grpc_server/requirements.txt
grpcio
grpcio-tools
python-dotenv
psycopg2-binary